Added dynamic post data on home

+++ routes/admin/posts.js
@@ -32,10 +32,10 @@ router.get('/edit/:id', (req, res) => {
 });
 
 router.post('/create/', (req, res) => {
-    let fileName = '';
+    let fileName = 'default.jpg';
     if (!isEmpty(req.files)){
         let file = req.files.picture; 
-        fileName = file.name;
+        fileName = Date.now() + '_' + file.name;
         file.mv('./public/uploads/' + fileName, err => {
             if(err) res.status(500).send(err);
         });
@@ -56,14 +56,16 @@ router.post('/create/', (req, res) => {
     });
     newPost.save().then(savedPost => {
         // console.log(savedPost);
+        req.flash('post_created', `${savedPost.title} was successfully saved as a ${savedPost.status} post.`);
         res.redirect('/admin/posts');
     }).catch(err =>{
-        // console.log(err);
+        console.log(err);
         res.status(500).send("Error saving the data");
     });   
 });
 
 router.put('/edit/:id', (req, res) => {
+    let fileName = 'default.jpg';
     let allowComments = true;
     if (req.body.allowComments){
         allowComments = true;
@@ -71,24 +73,54 @@ router.put('/edit/:id', (req, res) => {
     else{
         allowComments = false;
     }
-    Post.findByIdAndUpdate(req.params.id, {title: req.body.title, status: req.body.status, allowComments: allowComments, body: req.body.body})
-    .then(results => {
-        res.redirect('/admin/posts');
+    Post.findById(req.params.id).then(results =>{
+        results.title = req.body.title;
+        results.status = req.body.status;
+        results.allowComments = allowComments;
+        if (!isEmpty(req.files)){
+            let file = req.files.picture; 
+            fileName = Date.now() + '_' + file.name;
+            file.mv('./public/uploads/' + fileName, err => {
+                if(err) res.status(500).send(err);
+            });
+            results.files = fileName;
+        }
+        results.body = req.body.body;
+        results.save().then(updatedPost => {
+            req.flash('post_updated', `${updatedPost.title} was successfully updated!`);
+            res.redirect('/admin/posts');
+        }).catch(err => {
+            res.status(500).send(err);
+        });
     }).catch(err => {
-        res.status(500).send("Error Updated the Post");
+        console.log(err);
+        res.status(500).send("Error updating the post");
     });
+    // Post.findByIdAndUpdate(req.params.id, {title: req.body.title, status: req.body.status, allowComments: allowComments, body: req.body.body})
+    // .then(results => {
+    //     res.redirect('/admin/posts');
+    // }).catch(err => {
+    //     res.status(500).send("Error Updated the Post");
+    // });
 });
 
 router.delete('/:id', (req, res) => {
     Post.findById(req.params.id)
     .then(results => {
         if (results){
-            fs.unlink(uploadDir + results.files, (err) => {
-                if (err) throw err;
-            });
-            results.remove().then(obj => {res.redirect('/admin/posts')})
+            if (results.files === "default.jpg"){
+                
+            }
+            else{
+                fs.unlink(uploadDir + results.files, (err) => {
+                    if (err) throw err;
+                });
+            }
+            results.remove().then(obj => {
+                req.flash('post_deleted', `${results.title} was successfully deleted`);
+                res.redirect('/admin/posts');
+            })
             .catch(err => {console.log(err)});
-            
         }
     }).catch(err => {
         if (err) {
